adds a `uri` config option
The existing system of trying to intelligently generate a libvirt URI
from the user's config options was too restrictive for my uses. I have
added a `uri` config option that allows the user to directly specify the
libvirt resource they wish to connect to. This makes using
vagrant-libvirt with existing practices easier, and allows connecting
successfully in more environments (such as those that use
qemu:///session)