Support --no-destroy-on-error option
Vagrant supports a --no-destroy-on-error option to up to skip destroying
of the machine if there was an error on bringing it up. This can be
useful where an environment can trigger a bug which would normally
result in the domain being torn down preventing additional analysis.

Make sure to simply exit the loops by returning terminate, instead of
looking to execute the remaining retries.

Add spec tests to check that terminate does not call the runner to
remove the domain if the user has disabled destroy on error.

Define a missing constant for vagrant < 1.6 exposed by the added tests.