Bring up management network if it's not started yet
If I suspend using managedsave suspend_mode, then reboot the hypervisor,
and try to resume the box, then it fails because the management network
is not started (of course unless it's configured to autostart, which is
not the default behaviour).

This patch makes vagrant to start (and even recreate) the management
network if it's down.

Fixes vagrant-libvirt/vagrant-libvirt#761