check for 'ssh' service to see if interface is up
the 'echo' service is often not available out-of-the-box,
so the 'ping' of the ssh interface fails. This results in the
wrong ip addresses being used in /etc/exports, which results in
the vm not being able to mount the vagrant shares.

The 'vagrant ssh interface' will have the service 'ssh' available
(you won't be able to login to your vagrant box else), so it makes
sense to use that service for the ping.

This fixes the vm not being able to mount the shares over nfs after
a reload due to the wrong ip addresses being used.