Provide dedicated disk resolver class for devices
Create a dedicated disk resolver class to support resolving devices for
disks after resolving devices for the box volumes.

This is intended to have the resolving handled by a single specialist
class which can be called to give an approximation during validation to
check for duplicate definitions of the device attribute.

It may be expanded in the future.