Ensure sessions trigger a listing of the network devices
List network devices to trigger a side effect to ensure the
qemu-bridge-helper will run correctly for users of the modular daemons.

It is likely that this should be done for sessions to support resolving
the correct virtual bridge device anyway, however that is more involved
and maybe superseded by not needing a network interface should qemu
agent be able to act as a communicator. Or to perform port forwarding
using qemuargs to allow SSH access directly in.

Fixes: #1258